Add PermissionsImporter class for importing permissions from JSON files
